+++ CHANGELOG.md
@@ -105,7 +105,7 @@ Better state management, invite codes
 
 ### v0.4.0
 Invite codes
-- Client can get invite codes for document, persisted in db
+- Client can get invite codes for a document, persisted in db
 
 ### v0.4.1
 Better state management